requirements.txt
azure-common==1.1.23
azure-mgmt-compute==12.0.0
azure-identity==1.3.1
azure-mgmt-resource==10.1.0